Split docker dev and prod build stages
This patch makes the Dockerfile build both dev and prod images.  If build with
`docker build` it will default to building the prod stage, which gives us a
place to put development-specific build dependencies.

The `dev` stage is mostly an example at the moment: it installs
`build-essential` which I wanted for some local `make` tasks, but we may want
to add to it in future.

One important difference between the `dev` and `prod` stages is that the `prod`
stage is configured to run the `gunicorn` server via the script in
`management_interface/bin/serve`.  I've tweaked the workers and threads to be
somewhat sensible values, and for the logs to go to stdout.

As long as you're running it through `docker-compose`, the `dev` stage will
still run the `django` development server rather than `gunicorn`.